+++ tests/api/test_auth.py
@@ -20,26 +20,5 @@ async def test_auth_get_failure_with_headers(api_client_with_auth):
 async def test_auth_get_failure_without_headers(api_client):
     response = await api_client.get("/auth")
 
-    assert response.status_code == 422
-    assert response.json() == {
-        "detail": [
-            {
-                "type": "missing",
-                "loc": ["header", "nexus-endpoint"],
-                "msg": "Field required",
-                "input": None,
-            },
-            {
-                "type": "missing",
-                "loc": ["header", "nexus-bucket"],
-                "msg": "Field required",
-                "input": None,
-            },
-            {
-                "type": "missing",
-                "loc": ["header", "nexus-token"],
-                "msg": "Field required",
-                "input": None,
-            },
-        ]
-    }
+    assert response.status_code == 401
+    assert response.json() == {"message": "Unauthorized"}